+++ src/api/CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [0.30.2] - 2025-12-17
+
 ### Changed
 
 - Update the list of active operational units
@@ -639,7 +641,8 @@ update` command
 
 - Implement base FastAPI app
 
-[unreleased]: https://github.com/MTES-MCT/qualicharge/compare/v0.30.1...main
+[unreleased]: https://github.com/MTES-MCT/qualicharge/compare/v0.30.2...main
+[0.30.2]: https://github.com/MTES-MCT/qualicharge/compare/v0.30.1...v0.30.2
 [0.30.1]: https://github.com/MTES-MCT/qualicharge/compare/v0.30.0...v0.30.1
 [0.30.0]: https://github.com/MTES-MCT/qualicharge/compare/v0.29.0...v0.30.0
 [0.29.0]: https://github.com/MTES-MCT/qualicharge/compare/v0.28.0...v0.29.0